move classes in by.naxa.testdome package; add Gradle for build and dependency management.
