+++ README.md
@@ -4,6 +4,27 @@ Since I have been out of school for a few months, I figured that I could use som
 
 In this repository I am answering some sample Java interview questions supplied by testdome.com [(link to full set of questions)](https://www.testdome.com/d/java-interview-questions/4).
 
-Each file in this repository corresponds to one of the questions from the link above. While I assume the filenames will be descriptive enough to know which files correspond to which questions, I will also provide a direct link to the question I am answering in the header comments of the file. 
+## Project Structure
 
-All code was written and tested in Eclipse IDE for Java Developers, Version Neon Release (4.6.0)
+* src/main/java - content directory for application source code
+* src/test/java - content directory for tests source code
+
+Package `by.naxa.testdome` contains a set of files.
+
+Each file in this package corresponds to one of the questions from the link above. While I assume the filenames will be descriptive enough to know which files correspond to which questions, I will also provide a direct link to the question I am answering in the header comments of the file. 
+
+## Tools
+
+Initially, all code was written and tested in Eclipse IDE for Java Developers, Version Neon Release (4.6.0)
+
+Now the code is tested in IntelliJ IDEA (version 2019.1).
+
+Build system: Gradle.
+
+## Build
+
+Using Gradle:
+
+```bash
+gradlew build
+```
