+++ gradle/wrapper/gradle-wrapper.properties
@@ -1,0 +1,6 @@
+#Thu Feb 14 16:26:40 MSK 2019
+distributionUrl=https\://services.gradle.org/distributions/gradle-5.2-all.zip
+distributionBase=GRADLE_USER_HOME
+distributionPath=wrapper/dists
+zipStorePath=wrapper/dists
+zipStoreBase=GRADLE_USER_HOME
